+++ releaf-core/app/lib/releaf/core/error_formatter.rb
@@ -1,4 +1,4 @@
-module Releaf
+module Releaf::Core
   class ErrorFormatter
     attr_reader :errors